Refactor the factory methods
